refactor:application layer package 정리
