+++ app/src/main/java/com/codesoom/assignment/application/authentication/AuthenticationService.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.application;
+package com.codesoom.assignment.application.authentication;
 
 import com.codesoom.assignment.domain.User;
 import com.codesoom.assignment.domain.UserRepository;
